requirements.txt
flask
requests
beautifulsoup4
pypdf
python-dotenv
openai
textblob
spacy
langdetect
python-docx
python-pptx
openpyxl
pytesseract
pillow
chromadb
sentence-transformers
tiktoken
flask-login
flask-bcrypt
flask-sqlalchemy
email-validator
reportlab
markdown2
google-generativeai
sentence-transformers